Skip associating 'openlineage_events_completes' to 'ti' in AF3 (#1662)
Airflow 3 raises an error when attempting to associate
openlineage_events_completes with the task instance, saying that the
attribute does not exist on `ti`. The intended functionality to support
the requirements is expected to be worked around while addressing issue
#1635.

In the meantime, to avoid compatibility issues, this block of code is
conditionally skipped when running on Airflow 3.